refactor(runtime-vapor): split fragment.ts into domain modules
Pure code motion, no behavior change. fragment.ts had grown into a
1372-line integrator mixing the core branch pipeline with slot
boundary ambient state, the slot fallback state machine, and
hydration slot-boundary cursor state. Split by domain:

- slotBoundary.ts: SlotBoundaryContext + currentSlotBoundary ambient
  state, trackSlotBoundaryDirtying, hasSlotFallback (leaf module)
- slotFragment.ts: the SlotFallbackState machine (render/commit/
  recheck/dispose) + detachBlock + getRedirectedBoundary
- dom/hydrateFragment.ts: hydrating slot boundary cursor state
  (withHydratingSlotBoundary, getCurrentSlotEndAnchor,
  isHydratingSlotFallbackActive et al), which is hydration-cursor
  state shared by apiCreateFor/vdomInterop rather than slot logic
- EMPTY_BLOCK moves to block.ts next to the Block types

The SlotFragment class itself stays in fragment.ts: `extends
DynamicFragment` reads the base class binding at module evaluation
time, and componentSlots.ts (which constructs SlotFragment) sits
inside fragment.ts's evaluation closure, so hoisting the class into
another module can hit the base class before initialization
depending on entry order. The fallback machinery is plain function
declarations and therefore cycle-safe.

DynamicFragment.hydrate() and its helpers also stay for now:
CloseAnchorOwner is a const enum and isolatedModules forbids
cross-module const enum usage; they move together in a follow-up.

index.ts export surface is unchanged.
